trossen_slate -> libtrossen_slate, pytrossen_slate -> trossen_slate
