+++ cmake/version.hpp.in
@@ -26,11 +26,11 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-#ifndef TROSSEN_SLATE__VERSION_HPP_
-#define TROSSEN_SLATE__VERSION_HPP_
+#ifndef LIBTROSSEN_SLATE__VERSION_HPP_
+#define LIBTROSSEN_SLATE__VERSION_HPP_
 
-#define VERSION_MAJOR @trossen_slate_VERSION_MAJOR@
-#define VERSION_MINOR @trossen_slate_VERSION_MINOR@
-#define VERSION_PATCH @trossen_slate_VERSION_PATCH@
+#define VERSION_MAJOR @libtrossen_slate_VERSION_MAJOR@
+#define VERSION_MINOR @libtrossen_slate_VERSION_MINOR@
+#define VERSION_PATCH @libtrossen_slate_VERSION_PATCH@
 
-#endif  // TROSSEN_SLATE__VERSION_HPP_
+#endif  // LIBTROSSEN_SLATE__VERSION_HPP_
