+++ include/libtrossen_slate/base_driver.hpp
@@ -26,60 +26,45 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-#ifndef TROSSEN_SLATE__SERIAL_DRIVER_HPP_
-#define TROSSEN_SLATE__SERIAL_DRIVER_HPP_
+#ifndef LIBTROSSEN_SLATE__BASE_DRIVER_HPP_
+#define LIBTROSSEN_SLATE__BASE_DRIVER_HPP_
 
-#include <termio.h>
-
-#include <mutex>
 #include <string>
 
-typedef enum
+#include "serial_driver.hpp"
+
+namespace base_driver
 {
-  SYS_INIT = 0x00,
-  SYS_NORMAL,
-  SYS_REMOTE,
-  SYS_ESTOP,
-  SYS_CALIB,
-  SYS_TEST,
-  SYS_CHARGING,
 
-  SYS_ERR = 0x10,
-  SYS_ERR_ID,
-  SYS_ERR_COM,
-  SYS_ERR_ENC,
-  SYS_ERR_COLLISION,
-  SYS_ERR_LOW_VOLTAGE,
-  SYS_ERR_OVER_VOLTAGE,
-  SYS_ERR_OVER_CURRENT,
-  SYS_ERR_OVER_TEMP,
-
-  SYS_STATE_LEN,
-} SystemState;
-
-class SerialDriver
+typedef struct
 {
-public:
-  ~SerialDriver();
-  bool init(std::string & port, uint8_t id, int baud = B115200);
+  float cmd_vel_x;
+  float cmd_vel_y;
+  float cmd_vel_z;
+  uint32_t light_state;
 
-  bool getVersion(char * version);
-  bool setText(const char * text);
+  uint32_t system_state;
+  uint32_t charge;
+  float voltage;
+  float current;
+  float vel_x;
+  float vel_y;
+  float vel_z;
+  float odom_x;
+  float odom_y;
+  float odom_z;
+  uint32_t cmd;
+  uint32_t io;
+  uint32_t err;
+} ChassisData;
 
-  int readHoldingRegs(uint16_t addr, uint16_t cnt, uint16_t * data);
-  int writeHoldingRegs(uint16_t addr, uint16_t cnt, uint16_t * data);
-  int readWriteHoldingRegs(
-    uint16_t raddr, uint16_t rcnt, uint16_t * rdata,
-    uint16_t waddr, uint16_t wcnt, uint16_t * wdata);
+bool chassisInit(std::string & dev);
+bool getVersion(char * data);
+bool setText(const char * text);
+bool updateChassisInfo(ChassisData * data);
+bool setSysCmd(uint32_t cmd);
+bool setIo(uint32_t io);
 
-  int send(const void * data, int len, int timeout = 0);
-  int recv(uint8_t * data, int maxlen, int timeout = 0);
+} // namespace base_driver
 
-private:
-  void * m = nullptr;
-  int fd_ = -1;
-  fd_set read_set_;
-  std::mutex lock;
-};
-
-#endif // TROSSEN_SLATE__SERIAL_DRIVER_HPP_
+#endif // LIBTROSSEN_SLATE__BASE_DRIVER_HPP_
