+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>trossen_slate</name>
+  <name>libtrossen_slate</name>
   <version>1.0.0</version>
   <description>The trossen_slate package</description>
 
@@ -10,7 +10,7 @@
   <license>BSD-3-Clause</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
-  
+
   <export>
     <build_type>ament_cmake</build_type>
   </export>
