+++ python/trossen_slate/demo/basic_demo.py
@@ -26,44 +26,39 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-"""This demo showcases how to use more detailed methods to control and monitor the SLATE base."""
+"""This demo showcases how to use basic methods to control and monitor the SLATE base."""
 
-import trossen_slate as trossen
+import trossen_slate
 
 def main():
     # Create TrossenSlate object
-    slate = trossen.TrossenSlate()
+    slate = trossen_slate.TrossenSlate()
 
     # Initialize base and output result
-    success, result_init = slate.init_base()
-    print(f"Initialization success: {success}\nResult: {result_init}")
+    success, result = slate.init_base()
+    print(f"Initialization success: {success}\nResult: {result}")
 
-    # Disable charging and output result
-    success, result_charging = slate.enable_charging(False)
-    print(f"Disable charging success: {success}\nResult: {result_charging}")
-
-    # Enable motor torque and output result
-    success, result_torque = slate.enable_motor_torque(True)
-    print(f"Enable motor torque success: {success}\nResult: {result_torque}")
+    # Display "Hello world" on screen
+    slate.set_text("Hello world")
 
     while True:
-        # Initialize data with angular velocity and light state
-        my_data = trossen.ChassisData()
-        my_data.cmd_vel_z = -0.1
-        my_data.light_state = trossen.LightState.WHITE
+        # Set the LED colors to PURPLE
+        slate.set_light_state(trossen_slate.LightState.PURPLE)
 
-        # Write and update base data
-        slate.write(my_data)
+        # Set angular velocity to 0.1
+        slate.set_cmd_vel(0.0, 0.1)
 
-        # Initialize empty log data
-        log_data = trossen.ChassisData()
+        # Output charge percentage
+        print(f"Charge: {slate.get_charge()}%")
 
-        # Read and output data
-        slate.read(log_data)
+        # Get velocity and output values
+        vel = slate.get_vel()
+        print(f"Linear velocity: {vel[0]:.2f} Angular velocity: {vel[1]:.2f}")
 
-        print(f"Charge: {log_data.charge}%")
-        print(f"Linear velocity: {log_data.vel_x:.2f} Angular velocity: {log_data.vel_z:.2f}")
-        print(f"X: {log_data.odom_x:.2f} Y: {log_data.odom_y:.2f} Theta: {log_data.odom_z:.2f}")
+        # Get pose and output values
+        pose = slate.get_pose()
+        print(f"X: {pose[0]:.2f} Y: {pose[1]:.2f} Theta: {pose[2]:.2f}")
+
 
 if __name__ == "__main__":
     main()
