+++ requirements-dev.txt
@@ -1,0 +1,3 @@
+pybind11
+pybind11-stubgen
+scikit-build-core
