+++ src/base_driver.cpp
@@ -26,8 +26,8 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-#include "trossen_slate/base_driver.hpp"
-#include "trossen_slate/serial_driver.hpp"
+#include "libtrossen_slate/base_driver.hpp"
+#include "libtrossen_slate/serial_driver.hpp"
 
 namespace base_driver
 {
